Implementata Factory OggettiDiGioco per generare proiettili.
Risolto il bug #47 sulla creazione del Singleton OggettiDiGiocoFactory

Rimosso ServiceInviaAggiornamentiStato -> responsabilità riassegnata a ControllerGRASPFacade

Closes #41
Closes #42
Closes #43
Closes #40
Closes #44
Closes #45
Closes #47
