+++ src/main/java/it/univaq/dr2/tank19/model/collisione/Collisione.java
@@ -1,0 +1,4 @@
+package it.univaq.dr2.tank19.model.collisione;
+
+public interface Collisione {
+}
