--- src/main/java/it/univaq/dr2/tank19/service/ServiceInviaAggiornamentiStato.java
@@ -1,56 +1,0 @@
-package it.univaq.dr2.tank19.service;
-
-import it.univaq.dr2.tank19.model.TipoOggetto;
-import it.univaq.dr2.tank19.model.messaggi.MessaggioDiAggiornamentoStato;
-import it.univaq.dr2.tank19.model.oggettigioco.Proiettile;
-import org.springframework.messaging.simp.SimpMessagingTemplate;
-import org.springframework.scheduling.annotation.Scheduled;
-import org.springframework.stereotype.Service;
-
-/**
- * @author Carlo Centofanti
- * @created 09/06/2019
- */
-@Service
-public class ServiceInviaAggiornamentiStato {
-    private final ServiceTank serviceTank;
-    private final SimpMessagingTemplate simpMessagingTemplate;
-
-    public ServiceInviaAggiornamentiStato(ServiceTank serviceTank, SimpMessagingTemplate simpMessagingTemplate) {
-        this.serviceTank = serviceTank;
-        this.simpMessagingTemplate = simpMessagingTemplate;
-    }
-
-    @Scheduled(fixedDelay = 1000 / 60)
-    public void inviaAggiornamentiDiStato() {
-        serviceTank.findAll().iterator().forEachRemaining(tank -> {
-            String URLMessaggiPartita = "/partite/" + tank.getPartita().getId() + "/stato";
-            String direzione;
-            try {
-                direzione = tank.getDirezione().toString();
-            } catch (NullPointerException e) {
-                direzione = "null";
-            }
-
-            MessaggioDiAggiornamentoStato aggiornamentoTank = new MessaggioDiAggiornamentoStato();
-            MessaggioDiAggiornamentoStato aggiornamentoProiettile = new MessaggioDiAggiornamentoStato();
-
-            aggiornamentoTank.setIdOggetto(tank.getId());
-            aggiornamentoTank.setPosx(tank.getPosizione().getPosX());
-            aggiornamentoTank.setPosy(tank.getPosizione().getPosY());
-            aggiornamentoTank.setDirezione(direzione);
-            aggiornamentoTank.setTipoOggetto(TipoOggetto.CARRO_ARMATO);
-
-            if (tank.getProiettile() != null) {
-                Proiettile proiettile = tank.getProiettile();
-                aggiornamentoProiettile.setIdOggetto(proiettile.getId());
-                aggiornamentoProiettile.setPosx(proiettile.getPosizione().getPosX());
-                aggiornamentoProiettile.setPosy(proiettile.getPosizione().getPosY());
-                aggiornamentoProiettile.setDirezione(proiettile.getDirezione().toString());
-                aggiornamentoProiettile.setTipoOggetto(TipoOggetto.PROIETTILE);
-            }
-            simpMessagingTemplate.convertAndSend(URLMessaggiPartita, aggiornamentoTank);
-            simpMessagingTemplate.convertAndSend(URLMessaggiPartita, aggiornamentoProiettile);
-        });
-    }
-}
